pyproject.toml
[tool.poetry]
name = "mrc-backend"
version = "0.0.1"
description = ""
authors = ["Alban Tiberghien <atiberghien@numnprod.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
django-extensions = "^3.2.1"
django-cors-headers = "^3.14.0"
django-rest-framework = "^0.1.0"
django-allauth = "^0.52.0"
djangorestframework-simplejwt = "^5.2.2"
djangorestframework-camel-case = "^1.4.2"
django-filter = "^22.1"
drf-yasg = "^1.21.5"
django-autoslug = "^1.9.8"
django-taggit = "^3.1.0"
pillow = "^9.4.0"
drf-flex-fields = "^1.0.2"
djangorestframework-api-key = "^2.3.0"
django-environ = "^0.10.0"
dj-rest-auth = "^3.0.0"
paho-mqtt = "^1.6.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
